+++ R/create_project.R
@@ -15,12 +15,15 @@
 #' created.
 #' @param path_data A path. The directory where the secure data exist. Default is
 #' `NULL`.  When supplied, a symbolic link to data folder will be created.
-#' @param template Project template.
+#' @param template A project template. See
+#' \href{https://www.danieldsjoberg.com/starter/articles/create_project.html}{vignette}
+#' for details.
 #' @param git Logical indicating whether to create Git repository.  Default is `TRUE`
 #' When `NA`, user will be prompted whether to initialise Git repo.
 #' @param renv Logical indicating whether to add renv to a project.
-#' Default is `TRUE`. When `NA` which interactively asks user preference.
-#' @param overwrite Overwrite any existing files if they exist.  Options are
+#' Default is `TRUE`. When `NA` user is asked interactively for preference.
+#' @param overwrite Logical indicating whether to overwrite existing files
+#' if they exist.  Options are
 #' `TRUE`, `FALSE`, and `NA` (aka ask interactively).  Default is `NA`
 #' @param open Logical indicating whether to open new project in fresh RStudio
 #' session
@@ -79,8 +82,8 @@ create_project <- function(path, path_data = NULL, template = "default",
   # initializing renv project --------------------------------------------------
   if (isTRUE(renv)) {
     ui_done("Initialising {ui_field('renv')} project")
-    # set up sctruture of renv project
-    renv::scaffold(project = path)
+    # set up structure of renv project
+    renv::scaffold(project = path, settings = list(snapshot.type = "all"))
   }
 
   # if user added a path to a script, run it -----------------------------------
@@ -267,7 +270,7 @@ writing_files_folders <- function(selected_template, path,
       function(i) {
         if (!df_files$file_exists[i]) return(TRUE)
         if (isTRUE(overwrite)) return(TRUE)
-        if (!isTRUE(overwrite)) return(FALSE)
+        if (isFALSE(overwrite)) return(FALSE)
         if (!interactive()) return(FALSE)
         msg <- paste("{ui_path(df_files$filename[i])} already exists.",
                      "Would you like to overwrite it?")
@@ -341,4 +344,9 @@ copy_to_glue <- function(x) {
       purrr::list_modify(.x, glue = !.x$copy, copy = NULL)
     ) %||% .x
   )
+}
+
+# needed for R 3.4
+isFALSE = function(x) {
+  is.logical(x) && length(x) == 1L && !is.na(x) && !x
 }