+++ NEWS.md
@@ -1,3 +1,13 @@
+# starter 0.1.5
+
+* Default snapshot type is now 'all'.
+
+* Removed 'rstudioapi' package dependency.
+
+* Bug fix in `create_project(overwrite=)` where existing files were not being prompted with inquiry whether to overwrite.
+
+* Documentation updates.
+
 # starter 0.1.4
 
 * Documentation updates and tidying up for CRAN submission.